feat: Remote infrastructure provider: Refactor grpc connection creation to be shared between xds mutator and remote infrastructure
